backend/scripts/requirements.txt
Pillow==10.3.0
rembg==2.0.57
onnxruntime==1.17.1
